+++ RazorEx.DAL/Entities/Wallet.cs
@@ -16,7 +16,7 @@ namespace RazorEx.DAL.Entities
         [ForeignKey("WalletType")]
         public int TypeId { get; set; }
         public int UserId { get; set; }
-        public int Amount { get; set; }
+        public Int64 Amount { get; set; }
         public string Description { get; set; }
         public bool IsPay { get; set; }
